requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
distro==1.9.0
faiss-cpu==1.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
numpy==2.2.4
openai==1.70.0
packaging==24.2
pluggy==1.5.0
pydantic==2.11.2
pydantic_core==2.33.1
PyMuPDF==1.25.5
pytest==8.3.5
python-dotenv==1.1.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
mypy
fastapi
uvicorn[standard]
pydantic
rich